refactor: rename config submodule to axelar-config
