+++ packages/api/src/axelar-config.ts
@@ -1,0 +1,1 @@
+export * from "./axelar-config/index";
