--- packages/api/src/config.ts
@@ -1,1 +1,0 @@
-export * from "./config/index";
